Refactor statistics system to use Activity-based hierarchy.
Simplified the class structure for statistics by replacing the fragment-based approach with an activity-based hierarchy.

Changes include:
- Created abstract `BaseStatisticsActivity` to handle shared menus and lifecycle.
- Created intermediate `BaseStatisticsListActivity` for list-based statistics (Classic and Arcade).
- Implemented concrete `ClassicStatisticsActivity`, `ArcadeStatisticsActivity`, and `DailyStatisticsActivity`.
- Updated `MainActivity` and `BaseGameActivity` to use explicit intents for the new activities.
- Updated `AndroidManifest.xml` with the new activity definitions.
- Removed legacy `StatisticsActivity`, `BaseGameListFragment`, `BaseStatisticsFragment`, and specific fragment implementations.
- Updated unit tests (`NavigationTest`, `ColoringTest`) to reflect the new architecture.
- Verified that all tests pass.
