+++ app/src/main/java/com/antsapps/triples/BaseStatisticsActivity.java
@@ -1,0 +1,57 @@
+package com.antsapps.triples;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+import androidx.annotation.Nullable;
+import com.antsapps.triples.util.CsvExportable;
+
+public abstract class BaseStatisticsActivity extends BaseTriplesActivity implements CsvExportable {
+
+  @Override
+  protected void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+  }
+
+  @Override
+  public boolean onCreateOptionsMenu(Menu menu) {
+    MenuInflater inflater = getMenuInflater();
+    inflater.inflate(R.menu.statistics, menu);
+    return super.onCreateOptionsMenu(menu);
+  }
+
+  @Override
+  public boolean onOptionsItemSelected(MenuItem item) {
+    int itemId = item.getItemId();
+    if (itemId == android.R.id.home) {
+      finish();
+      return true;
+    } else if (itemId == R.id.help) {
+      Intent helpIntent = new Intent(getBaseContext(), HelpActivity.class);
+      startActivity(helpIntent);
+      return true;
+    } else if (itemId == R.id.settings) {
+      Intent settingsIntent = new Intent(getBaseContext(), SettingsActivity.class);
+      startActivity(settingsIntent);
+      return true;
+    } else if (itemId == R.id.export_to_csv) {
+      exportToCsv();
+      return true;
+    }
+    return super.onOptionsItemSelected(item);
+  }
+
+  @Override
+  public void onSignInSucceeded() {
+    super.onSignInSucceeded();
+    invalidateOptionsMenu();
+  }
+
+  @Override
+  public void onSignOut() {
+    super.onSignOut();
+    invalidateOptionsMenu();
+  }
+}
