+++ app/src/main/java/com/antsapps/triples/stats/DailyStatisticsActivity.java
@@ -1,4 +1,4 @@
-package com.antsapps.triples;
+package com.antsapps.triples.stats;
 
 import android.content.Context;
 import android.content.Intent;
@@ -18,7 +18,6 @@ import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.GestureDetector;
 import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
@@ -26,13 +25,15 @@ import android.widget.BaseAdapter;
 import android.widget.Button;
 import android.widget.GridView;
 import android.widget.TextView;
+import androidx.annotation.Nullable;
 import androidx.core.content.ContextCompat;
-import androidx.fragment.app.Fragment;
+import com.antsapps.triples.BaseStatisticsActivity;
+import com.antsapps.triples.DailyGameActivity;
+import com.antsapps.triples.R;
 import com.antsapps.triples.backend.Application;
 import com.antsapps.triples.backend.DailyGame;
 import com.antsapps.triples.backend.DailyStatisticsUtil;
 import com.antsapps.triples.backend.Game;
-import com.antsapps.triples.util.CsvExportable;
 import com.antsapps.triples.util.CsvUtil;
 import com.antsapps.triples.util.ShareUtil;
 import java.text.DateFormat;
@@ -47,7 +48,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
-public class DailyStatisticsFragment extends Fragment implements CsvExportable {
+public class DailyStatisticsActivity extends BaseStatisticsActivity {
 
   private Application mApplication;
   private List<DailyGame> mCompletedGames;
@@ -67,26 +68,32 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
   private TextView mDetailTime;
 
   @Override
-  public View onCreateView(
-      LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-    mApplication = Application.getInstance(getActivity());
-    View view = inflater.inflate(R.layout.daily_stats_fragment, container, false);
+  protected void onCreate(@Nullable Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    setContentView(R.layout.activity_daily_statistics);
+    setTitle(R.string.daily_label);
 
-    mMonthTitle = view.findViewById(R.id.month_title);
-    mCalendarGrid = view.findViewById(R.id.calendar_grid);
-    mCurrentStreakTv = view.findViewById(R.id.current_streak_tv);
-    mLongestStreakTv = view.findViewById(R.id.longest_streak_tv);
-    mTotalSolvedTv = view.findViewById(R.id.total_solved_tv);
-    mNextMonthBtn = view.findViewById(R.id.next_month);
+    if (getSupportActionBar() != null) {
+      getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+    }
 
-    mDetailDate = view.findViewById(R.id.detail_date);
-    mDetailStatus = view.findViewById(R.id.detail_status);
-    mDetailPlayBtn = view.findViewById(R.id.detail_play_btn);
+    mApplication = Application.getInstance(this);
+
+    mMonthTitle = findViewById(R.id.month_title);
+    mCalendarGrid = findViewById(R.id.calendar_grid);
+    mCurrentStreakTv = findViewById(R.id.current_streak_tv);
+    mLongestStreakTv = findViewById(R.id.longest_streak_tv);
+    mTotalSolvedTv = findViewById(R.id.total_solved_tv);
+    mNextMonthBtn = findViewById(R.id.next_month);
+
+    mDetailDate = findViewById(R.id.detail_date);
+    mDetailStatus = findViewById(R.id.detail_status);
+    mDetailPlayBtn = findViewById(R.id.detail_play_btn);
     mDetailPlayBtn.setBackgroundTintList(
-        ColorStateList.valueOf(ContextCompat.getColor(getActivity(), R.color.daily_accent)));
-    mDetailResultsContainer = view.findViewById(R.id.detail_results_container);
-    mDetailTriples = view.findViewById(R.id.detail_triples);
-    mDetailTime = view.findViewById(R.id.detail_time);
+        ColorStateList.valueOf(ContextCompat.getColor(this, R.color.daily_accent)));
+    mDetailResultsContainer = findViewById(R.id.detail_results_container);
+    mDetailTriples = findViewById(R.id.detail_triples);
+    mDetailTime = findViewById(R.id.detail_time);
 
     mCurrentMonth = Calendar.getInstance();
     mCurrentMonth.set(Calendar.DAY_OF_MONTH, 1);
@@ -95,7 +102,7 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
     mCurrentMonth.set(Calendar.SECOND, 0);
     mCurrentMonth.set(Calendar.MILLISECOND, 0);
 
-    view.findViewById(R.id.prev_month)
+    findViewById(R.id.prev_month)
         .setOnClickListener(
             v -> {
               mCurrentMonth.add(Calendar.MONTH, -1);
@@ -131,7 +138,7 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
 
     GestureDetector gestureDetector =
         new GestureDetector(
-            getActivity(),
+            this,
             new GestureDetector.SimpleOnGestureListener() {
               @Override
               public boolean onFling(
@@ -168,8 +175,6 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
 
     updateCalendar();
     updateStreaks();
-
-    return view;
   }
 
   private void updateCalendar() {
@@ -184,8 +189,7 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
     mNextMonthBtn.setEnabled(!isCurrentMonth);
 
     CalendarAdapter adapter =
-        new CalendarAdapter(
-            getActivity(), mCurrentMonth, mApplication.getDailyGames(), mSelectedDay);
+        new CalendarAdapter(this, mCurrentMonth, mApplication.getDailyGames(), mSelectedDay);
     mCalendarGrid.setAdapter(adapter);
 
     mCalendarGrid.setOnItemClickListener(
@@ -229,7 +233,7 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
       mDetailPlayBtn.setVisibility(View.VISIBLE);
       mDetailPlayBtn.setOnClickListener(
           v -> {
-            Intent intent = new Intent(getActivity(), DailyGameActivity.class);
+            Intent intent = new Intent(this, DailyGameActivity.class);
             DailyGame newGame = mApplication.getDailyGameForDate(mSelectedDay);
             intent.putExtra(Game.ID_TAG, newGame.getId());
             startActivity(intent);
@@ -248,9 +252,9 @@ public class DailyStatisticsFragment extends Fragment implements CsvExportable {
     }
   }
 
+  @Override
   public void exportToCsv() {
-    ShareUtil.shareCsv(
-        getActivity(), "daily_statistics.csv", CsvUtil.getDailyCsvContent(mCompletedGames));
+    ShareUtil.shareCsv(this, "daily_statistics.csv", CsvUtil.getDailyCsvContent(mCompletedGames));
   }
 
   private void updateStreaks() {
