--- app/src/main/res/layout/stats_fragment.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ListView xmlns:android="http://schemas.android.com/apk/res/android"
-  android:id="@android:id/list"
-  android:layout_width="fill_parent"
-  android:layout_height="fill_parent"
-  android:headerDividersEnabled="false" />
